LDAP example and more accurate volume mount hack
I think it would be useful to add an example ldap config. I had to do quite a bot of digging to find something that worked. A simple example can at least get a user pointed in the right direction. I also changed the volume mount hack to be a better example. The aiflow uid/gid is 50000, so it seems logical to have a sample that matches.