chore: Clarify the provenience and state of the chart
We have many users confusing the Official Apache Airlfow Community
chart with the User Community one, due to ambiguous "Community"
name used. We agreed to make the distinction of "User Community"
vs. "Official Apache Airflow Community" charts.

This PR fixes the ambiguity and also documents the current state
of both charts, so that the users are not confused about it.

Signed-off-by: Jarek Potiuk <jarek@potiuk.com>