fix: allow the use of the _CMD or _SECRET env vars
This allows the two alternate environment variables to be used for
setting the fernet and webserver secret keys. The priority support given
via the airflow.fernetKey and airflow.webserverSecretKey options
prevented the use of these since the variables being set have a higher
order of precedence when compared to the _CMD or _SECRET alternatives.

Signed-off-by: Alexander O'Donovan-Jones <alexodonovanjones@gmail.com>