feat: add support for .extraEnv (defaulting to []) for the various pod types
As per #629, add support for .extraEnv for the various pod types, including some of the validation and warnings.
This is generally better where we either use an environment variable that's something new, or that we are overriding
something in airflow.config (eg. setting the log level at DEBUG on workers but INFO everywhere else)

Signed-off-by: Matthew Byng-Maddick <mbyng-maddick@olx.com>